requirements.txt
Flask
gspread
oauth2client
python-telegram-bot==20.3